# server/controllers/EasyRecruitDemo_db/generated/CandidateSurveyControllerGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE CandidateSurveyControllerGenerated.js PLEASE EDIT ../CandidateSurveyController.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
import Properties from "../../../properties";
import CandidateSurveyModel from "../../../models/EasyRecruitDemo_db/CandidateSurveyModel";
import ErrorManager from "../../../classes/ErrorManager";
import { authorize } from "../../../security/SecurityManager";
import CandidateSurveyController from "../CandidateSurveyController";
const generatedControllers = {
  /**
   * Init routes
   */
  init: router => {
    const baseUrl = `${Properties.api}/candidatesurvey`;
    router.post(baseUrl + "", authorize([]), CandidateSurveyController.create);
    router.delete(baseUrl + "/:id", authorize([]), CandidateSurveyController.delete);
    router.get(baseUrl + "/findByCandidate/:key", authorize([]), CandidateSurveyController.findByCandidate);
    router.get(baseUrl + "/findBySurvey/:key", authorize([]), CandidateSurveyController.findBySurvey);
    router.get(baseUrl + "/:id", authorize([]), CandidateSurveyController.get);
    router.get(baseUrl + "", authorize([]), CandidateSurveyController.list);
    router.post(baseUrl + "/:id", authorize([]), CandidateSurveyController.update);
  },


  // CRUD METHODS


  /**
  * CandidateSurveyModel.create
  *   @description CRUD ACTION create
  *
  */
  create: async (req, res) => {
    try {
      const result = await CandidateSurveyModel.create(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * CandidateSurveyModel.delete
  *   @description CRUD ACTION delete
  *   @param ObjectId id Id
  *
  */
  delete: async (req, res) => {
    try {
      const result = await CandidateSurveyModel.delete(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * CandidateSurveyModel.findByCandidate
  *   @description CRUD ACTION findByCandidate
  *   @param Objectid key Id della risorsa Candidate da cercare
  *
  */
  findByCandidate: async (req, res) => {
    try {
      const result = await CandidateSurveyModel.findByCandidate(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * CandidateSurveyModel.findBySurvey
  *   @description CRUD ACTION findBySurvey
  *   @param Objectid key Id della risorsa Survey da cercare
  *
  */
  findBySurvey: async (req, res) => {
    try {
      const result = await CandidateSurveyModel.findBySurvey(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * CandidateSurveyModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id resource
  *
  */
  get: async (req, res) => {
    try {
      const result = await CandidateSurveyModel.get(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * CandidateSurveyModel.list
  *   @description CRUD ACTION list
  *
  */
  list: async (req, res) => {
    try {
      const result = await CandidateSurveyModel.list();
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  /**
  * CandidateSurveyModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id
  *
  */
  update: async (req, res) => {
    try {
      const result = await CandidateSurveyModel.update(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  // Custom APIs

};

export default generatedControllers;