# server/swagger.yaml
openapi: 3.0.0
info:
  title: easyRecruitDemo
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /candidate/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Candidate
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Candidate
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    post:
      summary: CRUD ACTION update
      tags:
        - Candidate
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Candidate"
...
  /notifications/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - CandidateNotification
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - CandidateNotification
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateNotification"
    post:
      summary: CRUD ACTION update
      tags:
        - CandidateNotification
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateNotification"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateNotification"
  /notifications/findByCandidate/{key}:
    get:
      summary: CRUD ACTION findByCandidate
      tags:
        - CandidateNotification
      parameters:
        - name: key
          in: path
          description: Id della risorsa Candidate da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CandidateNotification"
    
  #-----  CandidateSurvey  -----
...
  /candidatesurvey/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - CandidateSurvey
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - CandidateSurvey
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateSurvey"
    post:
      summary: CRUD ACTION update
      tags:
        - CandidateSurvey
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateSurvey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateSurvey"
  /candidatesurvey/findByCandidate/{key}:
    get:
      summary: CRUD ACTION findByCandidate
      tags:
        - CandidateSurvey
      parameters:
        - name: key
          in: path
          description: Id della risorsa Candidate da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CandidateSurvey"
  /candidatesurvey/findBySurvey/{key}:
    get:
      summary: CRUD ACTION findBySurvey
      tags:
        - CandidateSurvey
      parameters:
        - name: key
          in: path
          description: Id della risorsa Survey da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CandidateSurvey"
    
  #-----  Document  -----
...
  /document/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Document
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Document
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    post:
      summary: CRUD ACTION update
      tags:
        - Document
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Document"
  /document/findByDocument/{key}:
    get:
      summary: CRUD ACTION findByDocument
      tags:
        - Document
      parameters:
        - name: key
          in: path
          description: Id della risorsa Document da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    
  #-----  JobApplication  -----
...
  /survey/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Survey
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Survey
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
    post:
      summary: CRUD ACTION update
      tags:
        - Survey
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Survey"
    
  #-----  Task  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Candidate:
      properties:
        _id:
          type: string
        AppointmentDate:
          type: integer
        AppointmentId:
          type: string
        AppointmentResults:
          type: string
        AppointmentStatus:
          type: string
        ConversionFrom:
          type: string
        DeferReason:
          type: string
        DeferredDate:
          type: integer
        Email:
          type: string
        FirstName:
          type: string
        HireDate:
          type: integer
        Hired:
          type: boolean
        LastName:
          type: string
        MiddleName:
          type: string
        Recruiter:
          type: string
        ReferredBy:
          type: string
        SAPUser:
          type: string
      required:
        -  _id
    CandidateNotification:
      properties:
        _id:
          type: string
        CandidateId:
          type: string
        Email:
          type: string
        Message:
          type: string
        Method:
          type: string
        Phone:
          type: string
        SendDate:
          type: integer
        Status:
          type: string
        Title:
          type: string
        Type:
          type: string
      required:
        -  _id
    CandidateSurvey:
      properties:
        _id:
          type: string
        CandidateId:
          type: string
        Comments:
          type: string
        CompletionDate:
          type: integer
        Response:
          type: string
        Status:
          type: string
        SurveyId:
          type: string
        SurveyQuestions:
          type: string
      required:
        -  _id
    Document:
      properties:
        _id:
          type: string
        DocumentName:
          type: string
        DocumentType:
          type: string
        DocumentUrl:
          type: string
      required:
        -  _id
...
    Survey:
      properties:
        _id:
          type: string
        SurveyDescription:
          type: string
        SurveyName:
          type: string
        SurveyQuestions:
          type: string
      required:
        -  _id
...